requirements.txt
django
djangorestframework
pyymal
requests
django-cors-headers